+++ app-server/src/llm/mod.rs
@@ -146,6 +146,28 @@ pub(crate) enum ProviderClient {
 static ALWAYS_USE_REALTIME: OnceLock<bool> = OnceLock::new();
 const LLM_DEFAULT_HEADERS_JSON_ENV: &str = env::llm::DEFAULT_HEADERS_JSON;
 
+/// Whether the shared `LlmClient` actually initialized. Set from `main.rs`
+/// after client construction. Feature flags (e.g. `Feature::UserTaskExtraction`,
+/// `Feature::Signals`) only mirror the credential env vars, but `LlmClient::new`
+/// can still fail (bad `LLM_DEFAULT_HEADERS_JSON`, HTTP client build error, ...)
+/// — and when it does, the LLM-backed workers are never spawned, so enqueueing
+/// would strand messages on their queues unconsumed. Defaults to false so paths
+/// that never call `set_llm_client_available` (tests) don't enqueue.
+static LLM_CLIENT_AVAILABLE: OnceLock<bool> = OnceLock::new();
+
+/// Called once from `main.rs` right after `LlmClient` construction.
+/// First call wins (`OnceLock`); until then the LLM-backed producer hooks
+/// treat the client as unavailable and never enqueue.
+pub fn set_llm_client_available(available: bool) {
+    let _ = LLM_CLIENT_AVAILABLE.set(available);
+}
+
+/// Whether the shared `LlmClient` initialized. Every LLM-backed producer hook
+/// (user-task extraction, static-prompt extraction) gates on this.
+pub fn llm_client_available() -> bool {
+    LLM_CLIENT_AVAILABLE.get().copied().unwrap_or(false)
+}
+
 #[cfg_attr(not(feature = "signals"), allow(dead_code))]
 pub fn always_use_realtime() -> bool {
     *ALWAYS_USE_REALTIME.get().unwrap_or(&false)
